implement neural network (#13)
* Implement random and zero constructor for matrix

Random creates a (n x m) matrix with random values from -1.0 to 1.0.
Zero creates a (n x m) matrix with all values 0.0.

The constructors can be used to initialize a neural network more easily.

* Initialize neural network lib

* Implement funktionality for neural network

Neural network can be initialized with random weights between -1.0 -
1.0 and 0.0 biases.

The network can be used to calculate the output vector from the input
vector. This is done by calculating step by step the output vector for
each layer by weights[layer] * input_vector + biases[layer] =
output_vector. The calculated output vector is then used as input vector
for the next layer until all layers are processed.

* Fix formatting

* Fix formatting

* Implement error handling for neural network logic

Returning an error instead of panic to not crash the programm on error.

* Rename layers to layer_sizes

The name layer_sizes makes the meaning of this variable more clear.

* Renamed input values to input vector

* Add documentation comments

* Fix formatting and documentation test

* Implemnet usage of activation functions

For the outputs of the hidden layers Relu is used as activation
function. For the output layer Sigmoid is used. This decision was made
based on information from the internet and scientific papers.

---------

Co-authored-by: Jared Heinrich <jared.heinrich@sap.com>